refactor ChangeBlindnessUnitTest

+++ app/models/changeBlindness /Answer.java
@@ -12,8 +12,8 @@ public class Answer extends Model{
     @Id
     public long id;
     public double usedTime;
-    public int positionOfChangeX;
-    public int positionOfChangeY;
+    public double positionOfChangeX;
+    public double positionOfChangeY;
     public boolean isCorrect;
     @ManyToOne
     public User user;
